Move the fetching logic to the backend when resolving the user's session (during page refresh)
