+++ package.json
@@ -46,6 +46,7 @@
     "lodash.isempty": "4.4.0",
     "lodash.isequal": "4.5.0",
     "lodash.isobjectlike": "4.0.0",
+    "lodash.kebabcase": "4.1.1",
     "lodash.merge": "4.6.2",
     "lodash.now": "4.0.2",
     "lodash.remove": "4.7.0",
@@ -80,6 +81,7 @@
     "@types/lodash.includes": "4.3.6",
     "@types/lodash.isequal": "4.5.5",
     "@types/lodash.isobjectlike": "4.0.6",
+    "@types/lodash.kebabcase": "4.1.6",
     "@types/lodash.remove": "4.7.6",
     "@types/lodash.some": "4.6.6",
     "@types/lodash.sortby": "4.7.6",
