+++ src/gql/pages/index.ts
@@ -1,5 +1,10 @@
 import { gql } from 'graphql-request';
+import { Project } from '../../types/graphql/graphql';
 import project from '../fragments/project';
+
+export type IndexPageQueryResult = {
+  projects: Project[];
+}
 
 /**
  * Used by /src/pages/index.
