+++ src/pages/api/user.ts
@@ -3,7 +3,9 @@ import {
   NextApiResponse,
 } from 'next';
 import { getUserSession } from '../../lib/auth/userSession';
+import { UserModel } from '../../lib/faunadb/models/userModel';
 import { UserSession } from '../../types/auth/UserSession';
+import { Project } from '../../types/graphql/graphql';
 
 export type ApiGetUserResult = {
   user: UserSession | null;
@@ -18,6 +20,8 @@ type EndpointRequest = NextApiRequest & {
  *
  * Because the cookie containing the user session token can only be read by the server, we must use an API endpoint to retrieve it.
  *
+ * This endpoint is called every time a user refreshes the index page.
+ *
  * @param req
  * @param res
  */
@@ -25,11 +29,17 @@ export const user = async (req: EndpointRequest, res: NextApiResponse): Promise<
   const userSession: UserSession | undefined = await getUserSession(req);
 
   // The cookie contains the UserSession object
-  // It contains all the information we need for this POC, but if you want to fetch additional user-related and up-to-date data,
-  // you should do it here
   const result: ApiGetUserResult = {
     user: userSession || null,
   };
+
+  // If the user is authenticated, fetch their projects and add them in the user session
+  if (result?.user) {
+    const userModel = new UserModel();
+    const projects: Project[] = await userModel.getProjects(userSession as UserSession);
+
+    result.user.projects = projects;
+  }
 
   res.status(200).json(result);
 };
